Pull the metadata parsing logic into EventPipeMetadata class and stick in a separate file
